Add files via upload
In this project, I will be performing an unsupervised clustering of data on the customer's records from a company's database. Customer segmentation is the practice of separating customers into groups that reflect similarities among customers in each cluster. Customers are segmented to optimize the significance of each customer to the business, to modify products according to distinct needs and behaviour's of the customers. It also helps the business to finding concerns of different types of customers.